requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
colorthief>=0.2.1
opencv-python>=4.8.0
webcolors>=1.13
